+++ CHANGELOG.rst
@@ -15,7 +15,13 @@ Backwards incompatible changes
   The :class:`~tenpy.algorithms.dmrg.run` function works as befo
   In case you have directly used the :class:`~tenpy.algorithms.dmrg.EngineCombine` or :class:`~tenpy.algorithms.dmrg.EngineFracture`,
   you should update your code and use the :class:`~tenpy.algorithms.dmrg.TwoSiteEngine` instead.
+- Remove argument `leg0` from :class:`~tenpy.networks.mpo.MPOGraph.build_MPO`.
+- Remove argument `leg0` from :class:`~tenpy.networks.mpo.MPO.from_grids`, instead optionally give *all* `legs` as argument.
 
+Changed
+^^^^^^^
+- By default, for an usual MPO define `IdL` and `IdR` on all bonds. This can generate "dead ends" in the MPO graph of
+  finite systems, but it is useful for the `make_WI`/`make_WII` for MPO-exponentiation.
 
 Added
 ^^^^^
@@ -25,6 +31,7 @@ Added
 - Single-Site DMRG with the :class:`~tenpy.algorithms.dmrg.SingleSiteDMRG`.
 - Example function in ``examples/c_tebd.py`` how to run TEBD with a model originally having next-nearest neighbors.
 - :meth:`~tenpy.networks.mps.MPS.increase_L` to allow increasing the unit cell of an MPS.
+- Argument `insert_all_id` for :meth:`tenpy.networks.mpo.MPOGraph.from_terms` and :meth:`~tenpy.networks.mpo.MPOGraph.from_term_list`
 
 Fixed
 ^^^^^